+++ src/main/java/ticketsystem/DomainLayer/IRepository/IHistoryRepository.java
@@ -6,10 +6,10 @@ import ticketsystem.DomainLayer.history.Purchase;
 
 public interface IHistoryRepository {
     void addPurchase(Purchase purchase); 
-    Purchase findPurchaseById(int purchaseId);
+    Purchase findPurchaseById(long purchaseId);
     List<Purchase> getPurchasesByMemberId(long memberId);
-    List<Purchase> getPurchasesByCompanyId(int companyId);
+    List<Purchase> getPurchasesByCompanyId(long companyId);
     List<Purchase> getAllPurchases();
-    int generateNextId();
+    long generateNextId();
     
 }
